requirements.txt
folium
openai
Pyrebase4
python-dotenv
PyYAML
requests
streamlit
streamlit_authenticator
streamlit_folium
streamlit_lottie
